Organize files into folders and update require statements.
