+++ lib/actions/finder.rb
