+++ lib/db/db.rb
