+++ lib/db/reader.rb
