+++ lib/db/writer.rb
