+++ lib/ui/console.rb
