+++ spec/actions/lister_spec.rb
@@ -1,4 +1,4 @@
-require 'lister'
+require 'actions/lister'
 
 describe Lister do
